BEMF detection WIP

+++ src/sensors/Encoder.cpp
@@ -90,7 +90,8 @@ void Encoder::handleIndex() {
       // align encoder on each index
       long tmp = pulse_counter;
       // corrent the counter value
-      pulse_counter = round((double)pulse_counter/(double)cpr)*cpr;
+      // pulse_counter = round((double)pulse_counter/(double)cpr)*cpr;
+      pulse_counter = round(pulse_counter/cpr)*cpr;
       // preserve relative speed
       prev_pulse_counter += pulse_counter - tmp;
     }
